Make isEmailAvailable a custom command in the connector extension
