+++ web/app/connector-extension/custom-commands.js
@@ -1,3 +1,13 @@
-export const testCommand = () => (dispatch) => {
-    console.log('%cCustom Command Dispatched', 'color: green; font-size: x-large')
+import {makeJsonEncodedRequest} from 'progressive-web-sdk/dist/utils/fetch-utils'
+
+export const isEmailAvailable = (email) => (dispatch) => {
+    return makeJsonEncodedRequest(
+            '/rest/default/V1/customers/isEmailAvailable',
+            {customerEmail: email},
+            {method: 'POST'}
+        )
+        .then((response) => response.text())
+        .then((responseText) => {
+            return /true/.test(responseText)
+        })
 }
